Can't specify a non-".git" repository name
Sun Jul 18 08:34:51 2010: Request 59504 was acted upon.
Transaction: Ticket created by nomad@null.net
       Queue: Git-Repository
     Subject: Can't specify a non-".git" repository name
   Broken in: (no value)
    Severity: (no value)
       Owner: Nobody
  Requestors: nomad@null.net
      Status: new
 Ticket <URL: https://rt.cpan.org/Ticket/Display.html?id=59504 >

I'm trying to duplicate the following shell usage of Git (yes, I really
do not want to be using ".git" for the repo name, and have good reasons
for doing so):

    mkdir tmp$$
    cd tmp$$
    git --git-dir=.notgit --work-tree=. init

    echo -n "git-dir: "
    git --git-dir=.notgit --work-tree=. rev-parse --git-dir

    echo -n "is-bare-repository: "
    git --git-dir=.notgit --work-tree=. rev-parse --is-bare-repository

    cd ..
    rm -rf tmp$$

but this fails with Git::Repository, something to do with the
calculation of $wc_path. I have written some tests (attached) that show
the issue, but I'm afraid the logic in new() is a bit too unclear for
me to grok, in order to come up with my own fix.

Cheers,
Mark.
--
Mark Lawrence