Git::Repository: use better named git_dir and work_tree options and accessors
Initially, new() used 'repository' and 'working_copy' options,
and the accessors were 'repo_path' and 'wc_path'.

This poor naming scheme was inspired in part by Git.pm

This commit fixes by using names close to the git environment variables:
  GIT_DIR       => git_dir
  GIT_WORK_TREE => work_tree

The attributes updated by the options are being given identical names,
even if in some cases the resulting value might be different from the
provided option.

For backward compatibility, the old options and the old accessors
are being kept. When both the old and new options are provided to new(),
the new options silently take precedence.