Git::Repository: add support for mixins
The goal is to allow extensions to Git::Repository to be written,
without the need to add any code to Git::Repository. Collections
of methods can be put in their own namespace, and will be loaded
via import().

The idea came from discussions with Todd Rinaldo (TODDR) and
Aristotle Pagaltzis (ARISTOTLE). Thanks to them!