Update docs with example of options hash
I knew it was possible to specify a 'git' command but couldn't work out how from the docs. Had to read through the code to realise it had to go in a hashref, rather than just being part of the list of passed-in arguments. Hopefully this clarifies that there's an optional "options" hashref.