Tutorial entry "Ignore the system and global configuration files"
As documented in git-config(1), the environment variables
GIT_CONFIG_NOSYSTEM, XDG_CONFIG_HOME and HOME influence the
loading of system and global configuration files.

By setting those environment variables to specific values
whenever Git commands are run in the test repository, the
output of the commands is forced to the bare default.