Ensure $abs_git is always an absolute path
When the git binary is found in a $PATH component that is not absolute,
the 'real' symlink to $abs_git points to a destination relative to the
directory where 'real' lives, which is a temporary directory under t/,
i.e. two levels below the current working directory.

The 'real' symlink therefore resolves to something different than
expected, and _is_git() correctly claims that it is not a git binary.

This fixes RT #89086. (reported by Alex Raguero <jhavero@gmail.com>)