Add a new type of cache key, when the PATH has non-absolute components
When the PATH has non-absolute components, one is asking for trouble...

Since Git::Repository::Command chdir() into the repository before
exec()ing the git command, it's almost garanteed that the git that will
be run is not the expected one when doing things like:

    PATH=git-collection/1.8.3/bin:$PATH myscript.pl some/repository.git

because Git::Repository::Command will chdir() in some/, and from
there, the git binary will not be found under git-collection/ but
in the original PATH value.

Since the chdir()-before-exec() is supposed to be transparent, the cache
key for path-rel uses the current working directory and the PATH, and
instead of returning the initial name, _is_git() returns the absolute
path found by _which().