Add a library of Git utility functions, starting with cmp_git
I found myself copy-pasting the code for comparing Git versions in `sort`,
so I figured it belonged to a library.

I didn't create a version_cmp() method in Git::Repository, because if
I had, I would have given it a behaviour similar to that of the other
version_* methods, for consistency. They all call out to `git --version`
to get the version of the current Git binary. The resulting method would
have been impractical to use with sort().