Properly deal with versions that are references in the version_* methods
When $v was a Git::Version object, the `!ref` test would fail,
causing `( grep !ref, \@_ )[0]` to be an empty list. $v would end up
being the first option hash, if any.

This code explicitely picks the option hashes, and takes the first
of the remaining items to be the version number to test against.