Only drop "comments" when generating git versions based on the current one
This fixes failing test on a Mac OS X due to git from Xcode returning
a non-standard version.

The failing test was:

    ok( $r->version_eq($version), "$version version_eq $version" );

which was not comparing the Git version to itself anymore, because of the
mangling of $version done prior to generating known lesser and greater
versions to compare with the current one.