Add input_record_separator option to Git::Repository::Command::new
This option specifies a string which the $/ special variable is set to
during the command execution.

This allows one to read line by line the output of Git commands when
passing the -z option to them. One only needs to specify the "\0" string
as the input_record_separator.

This implements RT #134239.